runpod-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.3
aiofiles==23.2.1

# Computer Vision and ML
opencv-python-headless==4.9.0.80
opencv-contrib-python-headless==4.9.0.80
numpy==1.26.3
pillow==10.2.0

# TensorFlow and GPU acceleration
# Note: tensorflow (not tensorflow-gpu) includes GPU support since 2.13+
tensorflow==2.15.0
# tensorrt==8.6.1  # Already in CUDA base image, no need to install via pip
onnxruntime-gpu==1.17.0

# OCR and text detection
keras-ocr==0.9.3
pytesseract==0.3.10

# NSFW Detection
nudenet==3.4.2

# Audio processing and transcription
openai-whisper==20231117
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# Profanity filtering
better-profanity==0.7.0
profanity-check==1.0.3

# Video processing
ffmpeg-python==0.2.0
av==11.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
